requirements.txt
Pillow
matplotlib
rembg
pathos